fix(dracut): --remove allows to remove files from the host filesystem
The dracut temporary directory is under /var/tmp/dracut.???????/initramfs, so
using ".." specifiers in the path allows to escape from it.

E.g.:

```
$ mkdir /hostdir && ls -l /hostdir
total 0
$ dracut -f --remove "../../../../hostdir" test.img
$ ls -l /hostdir
ls: cannot access '/hostdir': No such file or directory
```

Fixes f8dfe3ee5b958262c38a821b15443893e400dba0

(cherry picked from commit 1abc4e0594aaeadf364774db9eb343d1ccd3572b)

+++ dracut.sh
@@ -3113,6 +3113,10 @@ fi
 
 # remove items
 for items in $remove_items; do
+    if [[ /$items/ == */../* ]]; then
+        derror "Refusing to remove '$items', '..' is not allowed."
+        continue
+    fi
     # shellcheck disable=SC2086
     rm -rf -- "${initdir:?}"/${items#/}
 done
